backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain-openai==0.1.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dateutil==2.8.2
pydantic>=2.7.4
python-multipart==0.0.9
